requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
certifi==2025.1.31
dill==0.4.0
flake8==7.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
platformdirs==4.3.7
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
pylint==3.3.6
python-dotenv==1.1.0
sniffio==1.3.1
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.2
